dinamic_b/app/requirements.txt
aiofiles==23.2.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
Babel==2.14.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
Django==3.2.16
djangorestframework==3.14.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.8.0
fastapi==0.110.1
fastapi-babel==0.0.9
FastAPI-SQLAlchemy==0.2.1
fastapi-users==13.0.0
flower==2.0.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
makefun==1.15.2
Mako==1.3.3
MarkupSafe==2.1.5
minio==7.2.5
numpy==1.26.4
packaging==24.0
passlib==1.7.4
pillow==10.3.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pwdlib==0.2.0
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pyotp==2.9.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
redis==5.0.3
rsa==4.9
sentry-sdk==1.45.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.4.4
starlette==0.37.2
text-unidecode==1.3
tomli==2.0.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
